test_requirements.txt
coverage==4.2
Django==1.10
PyYAML==3.11
